google_project_services is deprecated - use google_project_service (#4599)
