+++ google/resource_kms_crypto_key_test.go
@@ -371,16 +371,13 @@ resource "google_project" "acceptance" {
 	billing_account = "%s"
 }
 
-resource "google_project_services" "acceptance" {
+resource "google_project_service" "acceptance" {
 	project = "${google_project.acceptance.project_id}"
-
-	services = [
-	  "cloudkms.googleapis.com",
-	]
+	service = "cloudkms.googleapis.com"
 }
 
 resource "google_kms_key_ring" "key_ring" {
-	project  = "${google_project_services.acceptance.project}"
+	project  = "${google_project_service.acceptance.project}"
 	name     = "%s"
 	location = "us-central1"
 }
@@ -404,16 +401,13 @@ resource "google_project" "acceptance" {
 	billing_account = "%s"
 }
 
-resource "google_project_services" "acceptance" {
+resource "google_project_service" "acceptance" {
 	project = "${google_project.acceptance.project_id}"
-
-	services = [
-	  "cloudkms.googleapis.com",
-	]
+	service = "cloudkms.googleapis.com"
 }
 
 resource "google_kms_key_ring" "key_ring" {
-	project  = "${google_project_services.acceptance.project}"
+	project  = "${google_project_service.acceptance.project}"
 	name     = "%s"
 	location = "us-central1"
 }
@@ -435,16 +429,13 @@ resource "google_project" "acceptance" {
 	billing_account = "%s"
 }
 
-resource "google_project_services" "acceptance" {
+resource "google_project_service" "acceptance" {
 	project = "${google_project.acceptance.project_id}"
-
-	services = [
-	  "cloudkms.googleapis.com",
-	]
+	service = "cloudkms.googleapis.com"
 }
 
 resource "google_kms_key_ring" "key_ring" {
-	project  = "${google_project_services.acceptance.project}"
+	project  = "${google_project_service.acceptance.project}"
 	name     = "%s"
 	location = "us-central1"
 }
@@ -465,16 +456,13 @@ resource "google_project" "acceptance" {
 	billing_account = "%s"
 }
 
-resource "google_project_services" "acceptance" {
+resource "google_project_service" "acceptance" {
 	project = "${google_project.acceptance.project_id}"
-
-	services = [
-	  "cloudkms.googleapis.com",
-	]
+	service = "cloudkms.googleapis.com"
 }
 
 resource "google_kms_key_ring" "key_ring" {
-	project  = "${google_project_services.acceptance.project}"
+	project  = "${google_project_service.acceptance.project}"
 	name     = "%s"
 	location = "us-central1"
 }
@@ -500,16 +488,13 @@ resource "google_project" "acceptance" {
 	billing_account = "%s"
 }
 
-resource "google_project_services" "acceptance" {
+resource "google_project_service" "acceptance" {
 	project = "${google_project.acceptance.project_id}"
-
-	services = [
-	  "cloudkms.googleapis.com",
-	]
+	service = "cloudkms.googleapis.com"
 }
 
 resource "google_kms_key_ring" "key_ring" {
-	project  = "${google_project_services.acceptance.project}"
+	project  = "${google_project_service.acceptance.project}"
 	name     = "%s"
 	location = "us-central1"
 }
