+++ google/resource_kms_key_ring_test.go
@@ -129,15 +129,13 @@ resource "google_project" "acceptance" {
 	billing_account	= "%s"
 }
 
-resource "google_project_services" "acceptance" {
+resource "google_project_service" "acceptance" {
 	project  = "${google_project.acceptance.project_id}"
-	services = [
-		"cloudkms.googleapis.com"
-	]
+	service = "cloudkms.googleapis.com"
 }
 
 resource "google_kms_key_ring" "key_ring" {
-	project  = "${google_project_services.acceptance.project}"
+	project  = "${google_project_service.acceptance.project}"
 	name     = "%s"
 	location = "us-central1"
 }
@@ -153,11 +151,9 @@ resource "google_project" "acceptance" {
 	billing_account	= "%s"
 }
 
-resource "google_project_services" "acceptance" {
+resource "google_project_service" "acceptance" {
 	project  = "${google_project.acceptance.project_id}"
-	services = [
-		"cloudkms.googleapis.com"
-	]
+	service = "cloudkms.googleapis.com"
 }
 	`, projectId, projectId, projectOrg, projectBillingAccount)
 }
